support creation of AtkResNode and friends in 7.0
7.0 inlines the ctors of the child node types, so they are no longer available to sig. the ctors however are very simple so they've been replaced with manual implementations.

the only issue is that sigging the vtable addresses for the nodes requires large offsets, however the function changes incredibly rarely and these offsets will only require updating if they add a new node or the compiler version changes again.

+++ InteropGenerator.Runtime/Address.cs
@@ -6,10 +6,10 @@ public sealed class Address {
     public readonly ulong[] Mask;
 
     public readonly string Name;
-    public readonly byte[] RelativeFollowOffsets;
+    public readonly ushort[] RelativeFollowOffsets;
     public readonly string String;
     public nint Value;
-    public Address(string name, string @string, byte[] relativeFollowOffsets, ulong[] bytes, ulong[] mask, nint value) {
+    public Address(string name, string @string, ushort[] relativeFollowOffsets, ulong[] bytes, ulong[] mask, nint value) {
         Name = name;
         String = @string;
         RelativeFollowOffsets = relativeFollowOffsets;
